Git branch source doesn't support SCMSourceCriteria
