+++ blueocean-bitbucket-pipeline/src/main/java/io/jenkins/blueocean/blueocean_bitbucket_pipeline/BitbucketPipelineCreateRequest.java
@@ -6,6 +6,7 @@ import com.cloudbees.jenkins.plugins.bitbucket.BranchDiscoveryTrait;
 import com.cloudbees.plugins.credentials.common.StandardUsernamePasswordCredentials;
 import com.google.common.collect.Lists;
 import edu.umd.cs.findbugs.annotations.NonNull;
+import hudson.model.TaskListener;
 import io.jenkins.blueocean.commons.ErrorMessage;
 import io.jenkins.blueocean.commons.ServiceException;
 import io.jenkins.blueocean.credential.CredentialsUtils;
@@ -16,19 +17,27 @@ import io.jenkins.blueocean.scm.api.AbstractScmSourceEvent;
 import jenkins.branch.MultiBranchProject;
 import jenkins.model.Jenkins;
 import jenkins.model.JenkinsLocationConfiguration;
+import jenkins.scm.api.SCMHead;
+import jenkins.scm.api.SCMHeadObserver;
+import jenkins.scm.api.SCMRevision;
 import jenkins.scm.api.SCMSource;
 import jenkins.scm.api.SCMSourceOwner;
 import org.apache.commons.lang.StringUtils;
 import org.kohsuke.stapler.DataBoundConstructor;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
+import java.io.IOException;
 import java.util.List;
 
 /**
  * @author Vivek Pandey
  */
 public class BitbucketPipelineCreateRequest extends AbstractMultiBranchCreateRequest {
+    private static final Logger logger = LoggerFactory.getLogger(BitbucketPipelineCreateRequest.class);
+
     @DataBoundConstructor
     public BitbucketPipelineCreateRequest(String name, BlueScmConfig scmConfig) {
         super(name, scmConfig);
@@ -75,6 +84,30 @@ public class BitbucketPipelineCreateRequest extends AbstractMultiBranchCreateReq
             };
         }
         return null;
+    }
+
+    @Override
+    protected boolean repoHasJenkinsFile(@Nonnull SCMSource scmSource) {
+        final JenkinsfileCriteria criteria = new JenkinsfileCriteria();
+        try {
+            scmSource.fetch(criteria, new SCMHeadObserver() {
+                @Override
+                public void observe(@Nonnull SCMHead head, @Nonnull SCMRevision revision) throws IOException, InterruptedException {
+                    //do nothing
+                }
+
+                @Override
+                public boolean isObserving() {
+                    //if jenkinsfile is found stop observing
+                    return !criteria.isJekinsfileFound();
+
+                }
+            }, TaskListener.NULL);
+        } catch (IOException | InterruptedException e) {
+            logger.warn("Error detecting Jenkinsfile: "+e.getMessage(), e);
+        }
+
+        return criteria.isJekinsfileFound();
     }
 
     @Override
